requirements.txt
pydantic~=2.0
google-generativeai
pytest
flake8
autopep8
pyngrok
streamlit
PyPDF2
python-docx
pre-commit
